move test xDS server to its own library
