+++ build_autogenerated.yaml
@@ -8063,6 +8063,7 @@ targets:
   headers:
   - src/cpp/server/csds/csds.h
   - test/cpp/end2end/test_service_impl.h
+  - test/cpp/end2end/xds/xds_server.h
   src:
   - src/proto/grpc/testing/duplicate/echo_duplicate.proto
   - src/proto/grpc/testing/echo.proto
@@ -8100,6 +8101,7 @@ targets:
   - src/proto/grpc/testing/xds/v3/tls.proto
   - src/cpp/server/csds/csds.cc
   - test/cpp/end2end/test_service_impl.cc
+  - test/cpp/end2end/xds/xds_server.cc
   - test/cpp/end2end/xds_end2end_test.cc
   deps:
   - grpc++_test_config
